Refactor to pull common functionality into base types
1. `DiagnosticItem` and `CpsDiagnosticItem` are now derived from `BaseDiagnosticItem`.
2. `DiagnosticItemSource` and `CpsDiagnosticItemSource` are now derived from `BaseDiagnosticItemSource`.
3. `CpsDiagnosticItem.BrowseObject` and `CpsDiagnosticItemSource.DiagnosticDescriptorComparer` have been deleted, as these nested types now live in the respective base types.

The `AnalyzerItemTracker` and `AnalyzersCommandHandler` now operate on `BaseDiagnosticItem`s, which means the context menu now shows up properly. However, due to bugs (see #1919 and #1923) we can't yet use the context menu to change the severity.
