+++ src/VisualStudio/Core/SolutionExplorerShim/AnalyzersCommandHandler.cs
@@ -187,7 +187,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
 
         private bool ShouldShowDiagnosticContextMenu(IEnumerable<object> items)
         {
-            return items.All(item => item is DiagnosticItem);
+            return items.All(item => item is BaseDiagnosticItem);
         }
 
         private void UpdateDiagnosticContextMenu()
@@ -253,7 +253,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
 
             HashSet<ReportDiagnostic> selectedItemSeverities = new HashSet<ReportDiagnostic>();
 
-            var groups = _tracker.SelectedDiagnosticItems.GroupBy(item => item.AnalyzerItem.AnalyzersFolder.ProjectId);
+            var groups = _tracker.SelectedDiagnosticItems.GroupBy(item => item.ProjectId);
 
             foreach (var group in groups)
             {
@@ -412,7 +412,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
 
             foreach (var selectedDiagnostic in _tracker.SelectedDiagnosticItems)
             {
-                var projectId = selectedDiagnostic.AnalyzerItem.AnalyzersFolder.ProjectId;
+                var projectId = selectedDiagnostic.ProjectId;
                 var project = (AbstractProject)workspace.GetHostProject(projectId);
 
                 if (project == null)
