+++ src/VisualStudio/Core/SolutionExplorerShim/DiagnosticItem/DiagnosticItem.cs
@@ -15,59 +15,32 @@ using Roslyn.Utilities;
 
 namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplorer
 {
-    internal sealed partial class DiagnosticItem : BaseItem
+    internal sealed partial class DiagnosticItem : BaseDiagnosticItem
     {
-        private readonly DiagnosticDescriptor _descriptor;
-        private ReportDiagnostic _effectiveSeverity;
         private readonly AnalyzerItem _analyzerItem;
         private readonly IContextMenuController _contextMenuController;
 
-        public override event PropertyChangedEventHandler PropertyChanged;
-
         public DiagnosticItem(AnalyzerItem analyzerItem, DiagnosticDescriptor descriptor, ReportDiagnostic effectiveSeverity, IContextMenuController contextMenuController)
-            : base(string.Format("{0}: {1}", descriptor.Id, descriptor.Title))
+            : base(descriptor, effectiveSeverity)
         {
             _analyzerItem = analyzerItem;
-            _descriptor = descriptor;
-            _effectiveSeverity = effectiveSeverity;
+
             _contextMenuController = contextMenuController;
         }
 
-        public override ImageMoniker IconMoniker
+        protected override Workspace Workspace
         {
-            get
-            {
-                return MapEffectiveSeverityToIconMoniker(_effectiveSeverity);
-            }
+            get { return _analyzerItem.AnalyzersFolder.Workspace; }
         }
 
-        public AnalyzerItem AnalyzerItem
+        public override ProjectId ProjectId
         {
-            get
-            {
-                return _analyzerItem;
-            }
+            get { return _analyzerItem.AnalyzersFolder.ProjectId; }
         }
 
-        public DiagnosticDescriptor Descriptor
+        protected override AnalyzerReference AnalyzerReference
         {
-            get
-            {
-                return _descriptor;
-            }
-        }
-
-        public ReportDiagnostic EffectiveSeverity
-        {
-            get
-            {
-                return _effectiveSeverity;
-            }
-        }
-
-        public override object GetBrowseObject()
-        {
-            return new BrowseObject(this);
+            get { return _analyzerItem.AnalyzerReference; }
         }
 
         public override IContextMenuController ContextMenuController
@@ -76,74 +49,6 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
             {
                 return _contextMenuController;
             }
-        }
-
-        public Uri GetHelpLink()
-        {
-            if (BrowserHelper.TryGetUri(Descriptor.HelpLinkUri, out var link))
-            {
-                return link;
-            }
-
-            if (!string.IsNullOrWhiteSpace(Descriptor.Id))
-            {
-                _analyzerItem.AnalyzersFolder.Workspace.GetLanguageAndProjectType(_analyzerItem.AnalyzersFolder.ProjectId, out var language, out var projectType);
-
-                // we use message format here since we don't have actual instance of diagnostic here. 
-                // (which means we do not have a message)
-                return BrowserHelper.CreateBingQueryUri(Descriptor.Id, Descriptor.GetBingHelpMessage(), language, projectType);
-            }
-
-            return null;
-        }
-
-        internal void UpdateEffectiveSeverity(ReportDiagnostic newEffectiveSeverity)
-        {
-            if (_effectiveSeverity != newEffectiveSeverity)
-            {
-                _effectiveSeverity = newEffectiveSeverity;
-
-                NotifyPropertyChanged(nameof(EffectiveSeverity));
-                NotifyPropertyChanged(nameof(IconMoniker));
-            }
-        }
-
-        private ImageMoniker MapEffectiveSeverityToIconMoniker(ReportDiagnostic effectiveSeverity)
-        {
-            switch (effectiveSeverity)
-            {
-                case ReportDiagnostic.Error:
-                    return KnownMonikers.CodeErrorRule;
-                case ReportDiagnostic.Warn:
-                    return KnownMonikers.CodeWarningRule;
-                case ReportDiagnostic.Info:
-                    return KnownMonikers.CodeInformationRule;
-                case ReportDiagnostic.Hidden:
-                    return KnownMonikers.CodeHiddenRule;
-                case ReportDiagnostic.Suppress:
-                    return KnownMonikers.CodeSuppressedRule;
-                default:
-                    return default(ImageMoniker);
-            }
-        }
-
-        private void NotifyPropertyChanged(string propertyName)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-
-        internal void SetSeverity(ReportDiagnostic value, string pathToRuleSet)
-        {
-            UpdateRuleSetFile(pathToRuleSet, value);
-        }
-
-        private void UpdateRuleSetFile(string pathToRuleSet, ReportDiagnostic value)
-        {
-            var ruleSetDocument = XDocument.Load(pathToRuleSet);
-
-            ruleSetDocument.SetSeverity(_analyzerItem.AnalyzerReference.Display, _descriptor.Id, value);
-
-            ruleSetDocument.Save(pathToRuleSet);
         }
     }
 }
