+++ src/VisualStudio/Core/SolutionExplorerShim/SolutionExplorerShim.csproj
@@ -72,15 +72,15 @@
     <Compile Include="AnalyzersFolderItem\AnalyzersFolderItemProvider.cs" />
     <Compile Include="BaseItem.cs" />
     <Compile Include="BrowseObjectAttributes.cs" />
+    <Compile Include="DiagnosticItem\BaseDiagnosticItem.cs" />
+    <Compile Include="DiagnosticItem\BaseDiagnosticItemSource.cs" />
     <Compile Include="DiagnosticItem\CpsDiagnosticItem.cs" />
     <Compile Include="DiagnosticItem\CpsDiagnosticItemProvider.cs" />
     <Compile Include="DiagnosticItem\CpsDiagnosticItemSource.cs" />
-    <Compile Include="DiagnosticItem\CpsDiagnosticItem.BrowseObject .cs" />
     <Compile Include="DiagnosticItem\DiagnosticItem.BrowseObject.cs" />
     <Compile Include="DiagnosticItem\DiagnosticItemProvider.cs" />
     <Compile Include="AnalyzerReferenceManager.cs" />
     <Compile Include="ContextMenuController.cs" />
-    <Compile Include="DiagnosticItem\CpsDiagnosticItemSource.DiagnosticDescriptorComparer.cs" />
     <Compile Include="DiagnosticItem\DiagnosticItemSource.DiagnosticDescriptorComparer.cs" />
     <Compile Include="DiagnosticItem\RuleSelectionImage.cs" />
     <Compile Include="HierarchyItemToProjectIdMap.cs" />
